requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
alpha-vantage==2.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
backtrader==1.9.76.123
build==0.10.0
cachetools==4.2.2
ccxt==2.8.37
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
dateparser==1.1.7
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
Historic-Crypto==0.1.6
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
numpy==1.24.2
omegaconf==2.3.0
openai==0.11.3
openpyxl==3.1.1
packaging==23.0
pandas==1.3.5
pandas-stubs==1.5.3.230214
pip-tools==6.12.2
pycares==4.3.0
pycparser==2.21
pyproject_hooks==1.0.0
python-binance==1.0.16
python-dateutil==2.8.2
python-telegram-bot==13.7
pythonbot==0.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
TA-Lib==0.4.25
telegram==0.0.1
telegram-send==0.34
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
types-pytz==2022.7.1.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
websockets==10.4
yarl==1.8.2